Move logging to subdirectory
