+++ llarp/config/config.cpp
@@ -6,8 +6,8 @@
 #include <net/net.hpp>
 #include <router_contact.hpp>
 #include <util/fs.hpp>
-#include <util/logger_syslog.hpp>
-#include <util/logger.hpp>
+#include <util/logging/logger_syslog.hpp>
+#include <util/logging/logger.hpp>
 #include <util/mem.hpp>
 #include <util/memfn.hpp>
 #include <util/str.hpp>
