+++ llarp/messages/link_message_parser.cpp
@@ -9,7 +9,7 @@
 #include <messages/relay.hpp>
 #include <router_contact.hpp>
 #include <util/buffer.hpp>
-#include <util/logger.hpp>
+#include <util/logging/logger.hpp>
 #include <util/metrics.hpp>
 
 #include <memory>
