+++ llarp/messages/relay_status.cpp
@@ -7,7 +7,7 @@
 #include <routing/path_confirm_message.hpp>
 #include <util/bencode.hpp>
 #include <util/buffer.hpp>
-#include <util/logger.hpp>
+#include <util/logging/logger.hpp>
 #include <util/logic.hpp>
 #include <util/memfn.hpp>
