+++ llarp/util/fs.cpp
@@ -1,6 +1,6 @@
 #include <util/fs.hpp>
 
-#include <util/logger.hpp>
+#include <util/logging/logger.hpp>
 
 #include <fcntl.h>
 #include <sys/stat.h>
