+++ llarp/util/logging/ostream_logger.hpp
@@ -1,7 +1,7 @@
 #ifndef LLARP_UTIL_OSTREAM_LOGGER_HPP
 #define LLARP_UTIL_OSTREAM_LOGGER_HPP
 
-#include <util/logstream.hpp>
+#include <util/logging/logstream.hpp>
 #include <iostream>
 
 namespace llarp
