+++ llarp/util/threadpool.cpp
@@ -1,4 +1,4 @@
-#include <util/logger.hpp>
+#include <util/logging/logger.hpp>
 #include <util/time.hpp>
 #include <util/threadpool.h>
 #include <util/thread_pool.hpp>
